requirements.txt
graphics.py==5.0.1.post1
keyboard==0.13.2
